my_blog/requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==1.2.0
idna==2.9
jsonfield==3.1.0
Markdown==3.2.1
oauthlib==3.1.0
Pillow==7.0.0
Pygments==2.6.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.8
